+++ src/components/__tests__/FlawForm.spec.ts
@@ -1,3 +1,5 @@
+import { nextTick } from 'vue';
+
 import { describe, it, expect, vi } from 'vitest';
 import { DateTime } from 'luxon';
 import { flushPromises } from '@vue/test-utils';
@@ -12,6 +14,7 @@ import CweSelector from '@/components/CweSelector/CweSelector.vue';
 
 import { blankFlaw, useFlaw } from '@/composables/useFlaw';
 
+import DropDownMenu from '@/widgets/DropDownMenu/DropDownMenu.vue';
 import LabelEditable from '@/widgets/LabelEditable/LabelEditable.vue';
 import LabelDiv from '@/widgets/LabelDiv/LabelDiv.vue';
 import LabelSelect from '@/widgets/LabelSelect/LabelSelect.vue';
@@ -77,8 +80,7 @@ function mountWithProps(flaw: ZodFlawType, props: FlawFormProps) {
     },
     global: {
       stubs: {
-        // osimFormatDate not defined on test run, so we need to stub it
-        // EditableDate: true,
+        Teleport: true,
         RouterLink: true,
         AffectExpandableForm: true,
       },
@@ -355,29 +357,36 @@ describe('flawForm', () => {
 
   osimFullFlawTest('displays promote and reject buttons for state', async ({ flaw }) => {
     const subject = mountWithProps(flaw, { mode: 'edit' });
-    const workflowStateField = subject
-      .findAllComponents(IssueFieldState)
-      .find(component => component.text().includes('State'));
-    expect(
-      workflowStateField
-        ?.findAll('button')
-        ?.find(el => el.text() === 'Reject')
-        ?.text(),
-    ).toBe('Reject');
-    expect(
-      workflowStateField
-        ?.findAll('button')
-        ?.find(el => el.text().includes('Promote to'))
-        ?.text(),
-    ).toBe('Promote to Triage');
+    const menu = subject
+      .findComponent(IssueFieldState)
+      .findComponent(DropDownMenu);
+
+    const menuButton = menu.find('button[type="button"]');
+    await menuButton.trigger('click');
+    await nextTick();
+
+    for (const button of menu.findAll('button')) {
+      expect(([
+        'Change State',
+        'Promote to Triage',
+        'Reject',
+        'Reset',
+        'Revert',
+      ]).includes(button.text().trim())).toBe(true);
+    }
   });
 
   osimFullFlawTest('shows a modal for reject button clicks', async ({ flaw }) => {
     const subject = mountWithProps(flaw, { mode: 'edit' });
-    const workflowStateField = subject
-      .findAllComponents(IssueFieldState)
-      .find(component => component.text().includes('State'));
-    const rejectButton = workflowStateField?.findAll('button')?.find(el => el.text() === 'Reject');
+    const menu = subject
+      .findComponent(IssueFieldState)
+      .findComponent(DropDownMenu);
+
+    const menuButton = menu.find('button[type="button"]');
+    await menuButton.trigger('click');
+    await nextTick();
+
+    const rejectButton = menu?.findAll('button')?.find(el => el.text() === 'Reject');
     await rejectButton?.trigger('click');
     await flushPromises();
     expect(subject.find('.modal-dialog').exists()).toBe(true);